chore: set up initial project bones
So the architecture I am going for here will be a basic MVC because
I want to be able to unit test as much of the code base as possible.
The layers are as follows:

- app
  This is the "Controller" in the mvc work and is split further into:
  - routes
  - handlers
- model
  This will be pretty bare bones since the requirements are relatively
  simple from a business logic perspective.
- infra
  This is where external service/resource communication happens. The
  code here is not unit testtable so integration tests must be extensive
  to ensure no real regressions happen.
  The tests here do not negate the need for integration tests for the
  actual api tests as well.

In the interest of saving time things will be put into index.ts files
to start unless the line count goes super crazy.