add tests for the infra layer
As a stop gap I am leaving the test/jest configs where they are since I
only have a single 'resource' and there is no need to move things
around if they are currenty in a working state.

The infra layer is a wrapper around thee storage tech because I opted
to go with Postgres due to my familiarity with it.